Vähennetty duplikaatiota.
