+++ koodisto-api/src/main/java/fi/vm/sade/koodisto/service/types/CreateKoodistoDataType.java
@@ -65,5 +65,4 @@ public class CreateKoodistoDataType implements Serializable
         }
         return this.metadataList;
     }
-
 }
