+++ koodisto-api/src/main/java/fi/vm/sade/koodisto/service/types/common/KoodiType.java
@@ -3,10 +3,9 @@ package fi.vm.sade.koodisto.service.types.common;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
-public class KoodiType implements Serializable
+public class KoodiType extends KoodiAndKoodistoSharedType implements Serializable
 {
 
     private static final long serialVersionUID = 1L;
@@ -15,13 +14,9 @@ public class KoodiType implements Serializable
     protected KoodistoItemType koodisto;
     protected int versio;
     protected String koodiArvo;
-    protected Date paivitysPvm;
-    protected String paivittajaOid;
-    protected Date voimassaAlkuPvm;
-    protected Date voimassaLoppuPvm;
-    protected TilaType tila;
-    private List<KoodiMetadataType> metadata;
     protected long lockingVersion;
+    private List<KoodiMetadataType> metadata;
+
     public String getKoodiUri() {
         return koodiUri;
     }
@@ -62,44 +57,12 @@ public class KoodiType implements Serializable
         this.koodiArvo = value;
     }
 
-    public Date getPaivitysPvm() {
-        return paivitysPvm;
+    public long getLockingVersion() {
+        return lockingVersion;
     }
 
-    public void setPaivitysPvm(Date value) {
-        this.paivitysPvm = value;
-    }
-
-    public String getPaivittajaOid() {
-        return paivittajaOid;
-    }
-
-    public void setPaivittajaOid(String value) {
-        this.paivittajaOid = value;
-    }
-
-    public Date getVoimassaAlkuPvm() {
-        return voimassaAlkuPvm;
-    }
-
-    public void setVoimassaAlkuPvm(Date value) {
-        this.voimassaAlkuPvm = value;
-    }
-
-    public Date getVoimassaLoppuPvm() {
-        return voimassaLoppuPvm;
-    }
-
-    public void setVoimassaLoppuPvm(Date value) {
-        this.voimassaLoppuPvm = value;
-    }
-
-    public TilaType getTila() {
-        return tila;
-    }
-
-    public void setTila(TilaType value) {
-        this.tila = value;
+    public void setLockingVersion(long value) {
+        this.lockingVersion = value;
     }
 
     public List<KoodiMetadataType> getMetadata() {
@@ -107,13 +70,5 @@ public class KoodiType implements Serializable
             metadata = new ArrayList<>();
         }
         return this.metadata;
-    }
-
-    public long getLockingVersion() {
-        return lockingVersion;
-    }
-
-    public void setLockingVersion(long value) {
-        this.lockingVersion = value;
     }
 }
