+++ koodisto-api/src/main/java/fi/vm/sade/koodisto/service/types/common/KoodistoType.java
@@ -3,137 +3,43 @@ package fi.vm.sade.koodisto.service.types.common;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
-public class KoodistoType implements Serializable
+public class KoodistoType extends KoodiAndKoodistoSharedType implements Serializable
 {
 
     private static final long serialVersionUID = 1L;
     protected String koodistoUri;
     protected String resourceUri;
     protected int versio;
-    protected Date paivitysPvm;
-    protected String paivittajaOid;
-    protected Date voimassaAlkuPvm;
-    protected Date voimassaLoppuPvm;
-    protected TilaType tila;
     protected String omistaja;
     protected String organisaatioOid;
     protected Boolean lukittu;
-    private List<KoodistoMetadataType> metadataList;
     protected long lockingVersion;
+    private List<KoodistoMetadataType> metadataList;
 
-    /**
-     * Gets the value of the koodistoUri property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
     public String getKoodistoUri() {
         return koodistoUri;
     }
 
-    /**
-     * Sets the value of the koodistoUri property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
     public void setKoodistoUri(String value) {
         this.koodistoUri = value;
     }
 
-    /**
-     * Gets the value of the resourceUri property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
     public String getResourceUri() {
         return resourceUri;
     }
 
-    /**
-     * Sets the value of the resourceUri property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
     public void setResourceUri(String value) {
         this.resourceUri = value;
     }
 
-    /**
-     * Gets the value of the versio property.
-     * 
-     */
     public int getVersio() {
         return versio;
     }
 
-    /**
-     * Sets the value of the versio property.
-     * 
-     */
     public void setVersio(int value) {
         this.versio = value;
-    }
-
-    /**
-     * Gets the value of the paivitysPvm property.
-     * 
-     * @return
-     *     possible object is
-     *
-     *     
-     */
-    public Date getPaivitysPvm() {
-        return paivitysPvm;
-    }
-
-    public void setPaivitysPvm(Date value) {
-        this.paivitysPvm = value;
-    }
-
-    public String getPaivittajaOid() {
-        return paivittajaOid;
-    }
-
-    public void setPaivittajaOid(String value) {
-        this.paivittajaOid = value;
-    }
-
-    public Date getVoimassaAlkuPvm() {
-        return voimassaAlkuPvm;
-    }
-
-    public void setVoimassaAlkuPvm(Date value) {
-        this.voimassaAlkuPvm = value;
-    }
-
-    public Date getVoimassaLoppuPvm() {
-        return voimassaLoppuPvm;
-    }
-
-    public void setVoimassaLoppuPvm(Date value) {
-        this.voimassaLoppuPvm = value;
-    }
-
-    public TilaType getTila() {
-        return tila;
-    }
-
-    public void setTila(TilaType value) {
-        this.tila = value;
     }
 
     public String getOmistaja() {
@@ -160,13 +66,6 @@ public class KoodistoType implements Serializable
         this.lukittu = value;
     }
 
-    public List<KoodistoMetadataType> getMetadataList() {
-        if (metadataList == null) {
-            metadataList = new ArrayList<>();
-        }
-        return this.metadataList;
-    }
-
     public long getLockingVersion() {
         return lockingVersion;
     }
@@ -175,4 +74,10 @@ public class KoodistoType implements Serializable
         this.lockingVersion = value;
     }
 
+    public List<KoodistoMetadataType> getMetadataList() {
+        if (metadataList == null) {
+            metadataList = new ArrayList<>();
+        }
+        return this.metadataList;
+    }
 }
